requirements.txt
aiohttp==3.8.4
py-cord==2.4.1
python-dotenv==1.0.0